allocator: rename allocator2 to mma
`mma` standing for *Multi-Metric Allocator*. To avoid confusion with the
existing single metric allocator.

Informs: #103320
Release note: None
