+++ pkg/BUILD.bazel
@@ -227,9 +227,9 @@ ALL_TESTS = [
     "//pkg/kv/kvpb:kvpb_test",
     "//pkg/kv/kvprober:kvprober_test",
     "//pkg/kv/kvserver/abortspan:abortspan_test",
-    "//pkg/kv/kvserver/allocator/allocator2:allocator2_test",
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl_test",
     "//pkg/kv/kvserver/allocator/load:load_test",
+    "//pkg/kv/kvserver/allocator/mma:mma_test",
     "//pkg/kv/kvserver/allocator/plan:plan_test",
     "//pkg/kv/kvserver/allocator/storepool:storepool_test",
     "//pkg/kv/kvserver/apply:apply_test",
@@ -1427,12 +1427,12 @@ GO_TARGETS = [
     "//pkg/kv/kvprober:kvprober_test",
     "//pkg/kv/kvserver/abortspan:abortspan",
     "//pkg/kv/kvserver/abortspan:abortspan_test",
-    "//pkg/kv/kvserver/allocator/allocator2:allocator2",
-    "//pkg/kv/kvserver/allocator/allocator2:allocator2_test",
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl",
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl_test",
     "//pkg/kv/kvserver/allocator/load:load",
     "//pkg/kv/kvserver/allocator/load:load_test",
+    "//pkg/kv/kvserver/allocator/mma:mma",
+    "//pkg/kv/kvserver/allocator/mma:mma_test",
     "//pkg/kv/kvserver/allocator/plan:plan",
     "//pkg/kv/kvserver/allocator/plan:plan_test",
     "//pkg/kv/kvserver/allocator/storepool:storepool",
