+++ pkg/kv/kvserver/allocator/mma/allocator.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package allocator2
+package mma
 
 import "github.com/cockroachdb/cockroach/pkg/roachpb"
