+++ pkg/kv/kvserver/allocator/mma/testdata/means_memo
